XW-4232 | get rid of global functions
